Player Turn Function
Added the player turn function where the user can choose which dice to reroll and when they want to end their turn. This has future implementation possibilities for adding the underworld. The turn also automatically ends when the player rolls 3 or more skulls. The score counting at the end of their turn is currently being worked on.